requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
sounddevice==0.4.6
soundfile==0.12.1
cyp==0.10.1
numpy==1.24.3
flask-socketio==5.3.5